Minor fixes. The iostream include was missing in some examples.

+++ examples/quicksort/ff_qs_opt.cpp
@@ -138,7 +138,7 @@ void usage() {
 class my_loadbalancer: public ff_loadbalancer {
 protected:
     // implement your policy...
-    inline int selectworker() { return victim; }
+    inline size_t selectworker() { return victim; }
 
 public:
     // this is necessary because ff_loadbalancer has non default parameters....
@@ -147,7 +147,7 @@ public:
     void set_victim(int v) { victim=v;}
 
 private:
-    int victim;
+    size_t victim;
 };
